chore: move things to better dir structure
