+++ src/WPGraphQLPostsToPosts.php
@@ -23,11 +23,11 @@ final class WPGraphQLPostsToPosts {
 
 	private function create_instances() {
 		$this->instances['connections_registrar'] = new Connections\ConnectionsRegistrar();
-		$this->instances['post'] = new graphql\Types\Post();
-		$this->instances['users'] = new graphql\Types\Users();
-		$this->instances['post_mutation'] = new graphql\Mutations\Post();
-		$this->instances['users_mutation'] = new graphql\Mutations\Users();
-		$this->instances['field_types'] = new graphql\Fields();
+		$this->instances['field_types']           = new WPGraphQL\Types\Fields();
+		$this->instances['post']                  = new WPGraphQL\Types\Post();
+		// $this->instances['users']                 = new WPGraphQL\Types\Users();
+		// $this->instances['post_mutation'] = new graphql\Mutations\Post();
+		// $this->instances['users_mutation'] = new graphql\Mutations\Users();
 	}
 
 	private function register_hooks() {
